requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chromadb==0.3.29
click==8.1.7
clickhouse-connect==0.8.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
duckdb==1.1.3
emoji==2.14.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.85.1
ffmpy==0.4.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gradio==3.45.1
gradio_client==0.5.2
greenlet==3.1.1
h11==0.14.0
hnswlib==0.8.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.0.292
langdetect==1.0.9
langsmith==0.0.92
loguru==0.7.2
lxml==5.3.0
lz4==4.3.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.14.3
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numexpr==2.10.1
numpy==1.26.4
olefile==0.47
onnxruntime==1.19.2
openai==1.13.3
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
posthog==3.7.2
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pulsar-client==3.5.0
pure_eval==0.2.3
pycparser==2.22
pydantic==1.10.19
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.0
PyMuPDF==1.24.14
pyparsing==3.2.0
pypdf==5.1.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-iso639==2024.10.22
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.18
python-oxmsg==0.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
ruff==0.8.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.20.4
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.3
typer==0.14.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured @ file:///F:/Program/Python/Knowledge%20base-based%20assistant%20for%20large%20models/unstructured-0.16.8.tar.gz#sha256=33a70fb894cc1b703ade524707ab08b5e2d53eb7ca6258e42b24f9a4531b4952
unstructured-client==0.14.0
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
win32-setctime==1.1.0
wrapt==1.17.0
yarl==1.18.0
zhipuai==1.0.7
zipp==3.21.0
zstandard==0.23.0
